pl2 editor: moved state stuff into state.go
